--- handlers.h
@@ -1,38 +1,0 @@
-// Copyright 2023 Vahid Mardani
-/*
- * This file is part of chttpd.
- *  chttpd is free software: you can redistribute it and/or modify it under 
- *  the terms of the GNU General Public License as published by the Free 
- *  Software Foundation, either version 3 of the License, or (at your option) 
- *  any later version.
- *  
- *  chttpd is distributed in the hope that it will be useful, but WITHOUT ANY 
- *  WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS 
- *  FOR A PARTICULAR PURPOSE. See the GNU General Public License for more 
- *  details.
- *  
- *  You should have received a copy of the GNU General Public License along 
- *  with chttpd. If not, see <https://www.gnu.org/licenses/>. 
- *  
- *  Author: Vahid Mardani <vahid.mardani@gmail.com>
- */
-#ifndef HANDLERS_H_
-#define HANDLERS_H_
-
-
-#include <microhttpd.h>
-
-
-enum MHD_Result
-handle_hello(struct MHD_Connection *connection);
-
-
-enum MHD_Result
-handle_goodbye(struct MHD_Connection *connection);
-
-
-enum MHD_Result
-handle_echo(struct MHD_Connection *connection);
-
-
-#endif  // HANDLERS_H_
